Add files via upload
Scripts and data used for the calculation of extrapolated maps for cryptochrome. The "extrapolate_auto" script, which requires CCP4, python2.7 with NumPy and Matplotlib, as well as PHENIX, will use the MTZ files and the PDB files as input, and output extrapolated map coefficients for a range of assumed occupanies, which can be set in the script.